Creates new search type: Stop ID Search
Fixes #1264 - Stop search broken in Beta 18.1.0

Seems kind of redundant with with `StopID`, but we need another type and developers should only interface with this through a custom class. This commit routes all user-initiated searches for stop IDs through a new code path that does not assume a fully qualified internal stop ID.
